% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf.R
\name{as.data.frame.bnfClass}
\alias{as.data.frame.bnfClass}
\title{As Data Frame}
\usage{
\method{as.data.frame}{bnfClass}(x)
}
\arguments{
\item{x}{Object of class \code{bnfClass}, output of function \link{bnf}}
}
\value{
A data frame including:
\itemize{
\item time, where input y to function \link{bnf} is of class \code{ts}
\item y
\item trend
\item cycle
\item cycle_se
\item cycle_ci
}

If adjusted standard errors are applied
\itemize{
\item cycle_adjusted_se
\item cycle_ci_adjusted
}
}
\description{
Convert list object output to a data frame
}
