% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{audata}
\alias{audata}
\title{National Economic Indicators, Australia}
\format{
\subsection{\code{audata}}{

A data frame with 262 rows and 3 columns spanning: 1959-Q3 to 2024 Q4
\describe{
\item{date}{Observation quarter}
\item{UNRATE}{Australian Unemployment Rate, all persons seasonally adjusted,
Series ID: }
\item{GDPC1}{Australian Gross Domestic Product, chain volume measure
seasonally adjusted, $million, Series ID: }
}
}
}
\source{
\url{https://abs.gov.au}
}
\usage{
audata
}
\description{
National Economic Indicators, Australia
}
\keyword{datasets}
