% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf.R
\name{bnf}
\alias{bnf}
\title{Beveridge-Nelson Filter}
\usage{
bnf(
  y,
  delta_select = 2,
  fixed_delta = 0.01,
  p = 12,
  d0 = 0.01,
  dt = 5e-04,
  demean = c("nd", "sm", "pm", "dm"),
  iterative = 100,
  dynamic_bands = TRUE,
  adjusted_bands = FALSE,
  outliers = c(293, 294),
  window = 40,
  ib = TRUE,
  ...
)
}
\arguments{
\item{y}{vector, eg object of type \code{ts}}

\item{delta_select}{integer, set to 0 if use fixed delta, 1 if max amp-to-noise, 2 if min var(trend shocks)}

\item{fixed_delta}{integer, set a fixed delta to be used if delta_select=0, or for delta}

\item{p}{integer, is lag order for estimation}

\item{d0}{integer, min_delta: start point of grid search for delta if delta_select!=0}

\item{dt}{integer, delta increment}

\item{demean}{character, one of "nd", "sm", "dm", or "pm", where "nd" = no drift, "sm" = sample mean,
"dm" = dynamic demeaning, "pm" = structural breaks, set as 'breaks = c(100, 237)'}

\item{iterative}{integer, set to >1 for max number of iterations for iterative dynamic demeaning}

\item{dynamic_bands}{logical, set to TRUE for dynamic error bands, FALSE for fixed standard error bands}

\item{adjusted_bands}{logical, set to TRUE to adjusts for outlier observations when calculating error bands,}

\item{outliers}{set outliers as 'outliers = c(293, 294)'}

\item{window}{rolling window length for dynamic demeaning and/or dynamic error bands
(e.g., 40 is 10 years for quarterly data)}

\item{ib}{logical, set to FALSE if no iterative backcasting as in KMW2018 (just unconditional mean),
set to TRUE if iterative backcasting}

\item{...}{passed into error bands and piecewise_demean function}
}
\value{
List of class "bnf" containing:
\itemize{
\item call
\item y
\item cycle
\item trend
\item cycle_se: standard error
\item delta
\item demean_method
\item iterative
\item cycle_ci
\item cycle_adjusted_se
\item cycle_ci_adjusted
\item iterations
}
}
\description{
Based on Replication Files for Kamber, Morley & Wong
R conversion of some of the MATLAB codes originally written by Ben Wong
}
\examples{
data(usdata)
y <- transform_series(y = usdata$GDPC1, take_log = TRUE, pcode = "p1")
bnfOutput <- bnf(as.vector(y),
        delta_select = 2,
        demean = "dm",
        iterative = 100,
        dynamic_bands = TRUE,
        adjusted_bands = TRUE,
        outliers = c(293, 294),
        window = 40,
        ib = TRUE)
}
\references{
\itemize{
\item Kamber G, Morley J, Wong B (2018). “Intuitive and Reliable Estimates of the Output
Gap from a Beveridge-Nelson Filter.” The Review of Economics and Statistics, 100(3),
550-566. ISSN 0034-6535,.
\item Kamber G, Morley J, Wong B (2025). “Trend-cycle decomposition in the presence of
large shocks.” Journal of Economic Dynamics and Control, 173, 105066. ISSN 0165-1889.
}
}
