% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf.R
\name{piecewise_demean}
\alias{piecewise_demean}
\title{Piecewise Demean}
\usage{
piecewise_demean(y, breaks = c())
}
\arguments{
\item{y}{Numeric vector, a time series, eg of class ts}

\item{breaks}{Numeric vector, break points in observation indices, default \code{c()}
includes no breaks, then full sample mean is used instead}
}
\value{
Numeric vector, of length(y)
}
\description{
Demean relative to the break point(s) provided in 'breaks'.
The values in 'breaks' are assumed to be in ascending order.
}
\examples{
data(usdata)
piecewise_demean(usdata$GDPC1, breaks=c(100,200))
}
