% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf.R
\name{plot.bnfClass}
\alias{plot.bnfClass}
\title{Plot}
\usage{
\method{plot}{bnfClass}(
  x,
  main = "BN Filter Cycle",
  plot_ci = TRUE,
  col = "red",
  secondary_col = "blue",
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{object of class "bnf" output of \link{bnf}}

\item{main}{character, plot title}

\item{plot_ci}{logical, include cycle confidence interval}

\item{col}{character, primary color or hexcode}

\item{secondary_col}{character, secondary color or hexcode}

\item{lwd}{numeric, line width}

\item{...}{other arguments, not used}
}
\value{
A plot
}
\description{
Plot method for object of class "bnf"
}
\examples{
data(usdata)
y <- transform_series(y = usdata$GDPC1, take_log = TRUE, pcode = "p1")
bnfOutput <- bnf(as.vector(y),
        delta_select = 2,
        demean = "dm",
        iterative = 100,
        dynamic_bands = TRUE,
        adjusted_bands = TRUE,
        outliers = c(293, 294),
        window = 40,
        ib = TRUE)
plot(bnfOutput, main="US Output Gap")
}
