% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf.R
\name{print.bnfClass}
\alias{print.bnfClass}
\title{Print}
\usage{
\method{print}{bnfClass}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class bnf output of \link{bnf}}

\item{digits}{integer, significant digits in print, default \code{getOption("digits") - 3}}

\item{...}{further arguments, not used}
}
\value{
Nothing
}
\description{
Print method for class "bnf"
}
\examples{
data(usdata)
y <- transform_series(y = usdata$GDPC1, take_log = TRUE, pcode = "p1")
bnfOutput <- bnf(as.vector(y),
        delta_select = 2,
        demean = "dm",
        iterative = 100,
        dynamic_bands = TRUE,
        adjusted_bands = TRUE,
        outliers = c(293, 294),
        window = 40,
        ib = TRUE)
print(bnfOutput)
}
