% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnf.R
\name{transform_series}
\alias{transform_series}
\title{Transform Series}
\usage{
transform_series(
  y,
  take_log = FALSE,
  dcode = c("nd", "d1", "d4", "d12"),
  pcode = c("np", "p1", "p4", "p12")
)
}
\arguments{
\item{y}{numeric vector, input time series, eg class ts}

\item{take_log}{logical, take the natural logarithm? Default is no (FALSE)}

\item{dcode}{character option to specify how y is differenced:
\itemize{
\item "nd", No difference, i.e., level, default
\item "d1", 1st Difference, i.e., (1 - B)y
\item "d4", 4th Difference, i.e., (1 - B^4)y, use with quarterly data
\item "d12", 12th Difference, i.e., (1 - B^12)y, use with monthly data
}}

\item{pcode}{character, option to specify if percentages are computed:
\itemize{
\item "np", no change, default
\item "p1", multiply by 100
\item "p4", multiply by 400, eg. for annualized quarterly rate
\item "p12", multiply by 1200, eg. for annualized monthly rate
}}
}
\value{
A vector transformed, if class x is class ts, class ts
}
\description{
Returns a suitably transformed, lagged and/or differenced data.
Inspired by the GAUSS procs originally written by \href{http://www.princeton.edu/~mwatson/wp.html}{Mark Watson}.
}
\examples{
data(usdata)
head(transform_series(usdata$GDPC1, TRUE, "d4", "p4"))
}
