% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usdata}
\alias{usdata}
\title{National Economic Indicators, United States}
\format{
\subsection{\code{usdata}}{

A data frame with 278 rows and 5 columns spanning: 1947 Q1 to 2016 Q4. The original
dataset from GMW2018.
\describe{
\item{DATE}{Observation quarter}
\item{UNRATE}
\item{GDPC1}
\item{cpiaucsl}
\item{GDPDEF}
\item{UNRATE}
}
}
}
\source{
\url{https://bnfiltering.com/}
}
\usage{
usdata
}
\description{
National Economic Indicators, United States
}
\keyword{datasets}
