% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgdp}
\alias{usgdp}
\title{United States Gross Domestic Product}
\format{
\subsection{\code{usgdp}}{

A time series object of length 305 spanning 1947 Q1 to 2023 Q4
}
}
\source{
\url{https://bnfiltering.com/}
}
\usage{
usgdp
}
\description{
United States Gross Domestic Product
}
\keyword{datasets}
